[update] fixed bugs

+++ include/hare/base/log/async.h
@@ -17,7 +17,6 @@
 #include <hare/base/util/count_down_latch.h>
 
 #include <string>
-#include <thread>
 #include <vector>
 
 namespace hare {
@@ -28,7 +27,7 @@ namespace log {
         static const int32_t META_FLUSH_INTERVAL = 1;
         static const int64_t BLOCK_NUMBER = 16;
 
-        using fixed_block = io::fixed_cache<HARE_LARGE_BUFFER>;
+        using fixed_block = io::fixed_cache<static_cast<size_t>(HARE_LARGE_BUFFER)>;
         using blocks = std::vector<fixed_block::ptr>;
 
         std::mutex mutex_ {};
